// src/app/shared/components/models/pricing.model.ts
export class Pricing {
  location: any[] = ['fullfront'];
  garmentQuality: string = 'good';
  colors: number = 1;
  colorsCount: number;
  underbaseColor: boolean = true;
  pricePerShirt: number = 0;
  printPrice: number = 0;
  garmentPrice: number = 2.65;
  printCostAndColor: number = 0;
  quantity: number = 12;
  priceExTax: any;
  totalPrice: any;
}